Add pages for ciphers

+++ app/src/main/java/com/renan/cifraeditor/data/tables/Cipher.kt
@@ -6,8 +6,8 @@ import androidx.room.PrimaryKey
 
 @Entity
 data class Cipher(
-    @PrimaryKey val id: Int?=null,
+    @PrimaryKey val id: Long?=null,
     @ColumnInfo var name: String,
     @ColumnInfo var artist: String? = null,
-    @ColumnInfo(name = "fk_tom") var fkTom: Int
+    @ColumnInfo(name = "fk_tom") var fkTom: Long
 )
