+++ app/src/main/java/com/renan/cifraeditor/data/tables/Word.kt
@@ -9,5 +9,5 @@ data class Word(
    @PrimaryKey val id:Int?=null,
    @ColumnInfo  var name:String,
     @ColumnInfo(name = "fk_chord") val fkChord:Int?,
-   @ColumnInfo(name = "fk_chipher")  val fkChiper:Int
+   @ColumnInfo(name = "fk_chipher")  val fkChiper:Long
 )
